Revert "Update strace to 4.18."
This reverts commit dc75b01004a0588c1eb3bc26d7248a6e473b2cdd.

+++ linux/nios2/arch_regs.h
@@ -1,1 +1,1 @@
-extern unsigned int *const nios2_sp_ptr;
+extern unsigned long *const nios2_sp_ptr;
